+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.13] - Unreleased
 
+### Added
+
+- Registry class to manage Cycle classes.
+- Use of Concurrent::Set for thread safety.
+
 ## [0.1.12] - 2025-09-05
 
 ### Added